# Vectorize TrimTransparentPixels in GifEncoderCore