@gfoidl 2023-07-29
😃 
I think the easiest way to understand / check is to use the benchmark-code (see top-comment) in a simple console app and step with the debugger through it (maybe change the size of the `_rowSpan` to 20 or that like.
Calculation of the correct index for `end` is the strangest part IMO.