@JimBobSquarePants 2023-08-09
This is fantastic stuff. I figured theoretically after reading some of the source for `Span.IndexOf` that masking with bit counting would be the vectorized solution I just had no idea how I'd actually implement it.

Tip of the cap to you sir.